Renamed: component PickerAttributes.vue
