+++ apps/frontend/src/components/page-products/PickerAttributes.vue
